requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
passlib[bcrypt]
python-jose
